@@ -2,27 +2,32 @@
   "name": "eva",
   "version": "1.59.0",
   "description": "Evolution Vizualization Application",
-  "main": "index.js",
-  "scripts": {
-    "dev": "webpack-dev-server --watch-poll --mode development",
-    "build": "webpack --mode production",
-    "test": "jest  --config=jest.config.js --no-cache"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 3 version"
-  ],
   "author": "Fred",
   "license": "ISC",
+  "private": true,
+  "scripts": {
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "test:unit": "vue-cli-service test:unit",
+    "lint": "vue-cli-service lint",
+    "lint-no-fix": "vue-cli-service lint --no-fix"
+  },
   "devDependencies": {
     "@babel/core": "^7.6.2",
     "@babel/preset-env": "^7.6.2",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-router": "~4.5.0",
+    "@vue/cli-plugin-unit-mocha": "~4.5.0",
+    "@vue/cli-plugin-vuex": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
     "@vue/test-utils": "^1.0.0-beta.33",
     "autoprefixer": "^9.6.4",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.3.0",
     "babel-loader": "^8.0.6",
+    "chai": "^4.1.2",
     "css-loader": "^3.2.0",
     "css-mqpacker": "^7.0.0",
     "cssnano": "^4.1.10",
@@ -34,24 +39,26 @@
     "jest": "^25.3.0",
     "jest-fetch-mock": "^3.0.3",
     "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^4.14.1",
     "path": "^0.12.7",
     "postcss-loader": "^3.0.0",
     "regenerator-runtime": "^0.13.3",
-    "sass": "^1.23.0",
+    "sass": "^1.35.2",
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
+    "vue-cli-plugin-vuetify": "~2.4.1",
     "vue-jest": "^3.0.5",
-    "vue-loader": "^15.7.1",
-    "vue-style-loader": "^4.1.2",
+    "vue-loader": "^15.9.6",
+    "vue-style-loader": "^4.1.3",
     "vue-template-compiler": "^2.6.10",
-    "vuetify-loader": "^1.3.0",
-    "webpack": "^4.41.0",
-    "webpack-cli": "^3.3.11",
+    "vuetify-loader": "^1.7.2",
+    "webpack": "^4.46.0",
+    "webpack-cli": "^4.7.2",
     "webpack-dev-server": "^3.8.2"
   },
   "dependencies": {
     "@mdi/js": "^4.5.95",
+    "core-js": "^3.6.5",
     "d3": "^5.12.0",
     "file-saver": "^2.0.2",
     "jointjs": "^3.2.0",
@@ -59,7 +66,7 @@
     "leaflet.markercluster": "^1.5.0",
     "leaflet.tilelayer.colorfilter": "^1.2.5",
     "vue": "^2.6.10",
-    "vue-async-computed": "^3.7.0",
+    "vue-async-computed": "^3.9.0",
     "vue-ctk-date-time-picker": "^2.4.0",
     "vue-draggable-resizable": "^2.2.0",
     "vue-html2canvas": "0.0.4",
