@@ -3,43 +3,10 @@
     <div v-if="error" class="error-message">
       {{ error }}
     </div>
-    <!-- <div v-if="!error" class="wrapper-property">
-      <v-select
-        v-model="maptheme"
-        class="select-property"
-        :items="['default', 'black']"
-        label="Тема"
-        @change="changeMapTheme"
-      />
-      <v-select
-        v-model="clusterTextCount"
-        class="select-property"
-        :items="[3, 4, 5, 6]"
-        label="Количество элементов в строке подписей"
-        @change="changeClusterTextCount"
-      />
-      <v-text-field
-        v-model="clusterDelimiter"
-        label="Разделитель"
-        @blur="blurClusterDelimiter"
-        @keyup.enter="blurClusterDelimiter"
-      />
-      <v-select
-        v-model="clusterPosition"
-        class="select-property"
-        :items="clusterPositionItems"
-        item-text="name"
-        item-value="id"
-        chips
-        multiple
-        label="Порядок элементов"
-        @blur="blurClusterPosition"
-      />
-    </div> -->
     <div
-      ref="map"
       v-if="!error"
       id="mapContainer"
+      ref="map"
       :style="{
         width: `${Math.trunc(widthFrom)}px`,
         height: `${Math.trunc(heightFrom) - top / 2 - 45}px`,
@@ -57,7 +24,7 @@ import "leaflet.markercluster";
 import vuetify from "../../plugins/vuetify";
 import dashMapUserSettings from "./dashMapUserSettings.vue";
 import store from "../../store/index.js"; // подключаем файл с настройками хранилища Vuex
-
+import Vue from "vue";
 export default {
   props: {
     // переменные полученные от родителя
@@ -69,9 +36,6 @@ export default {
     widthFrom: null, // ширина родительского компонента
     heightFrom: null, // выоста родительского компонента
     options: Object,
-  },
-  components: {
-    dashMapUserSettings,
   },
   data() {
     return {
@@ -122,28 +86,22 @@ export default {
     this.initClusterDelimiter();
     const unsubscribe = store.subscribe((mutation, state) => {
       if (mutation.type == "updateOptions") {
-        this.map.setView(this.startingPoint, mutation.payload.options.zoomLevel);
-        this.map.wheelPxPerZoomLevel = mutation.payload.options.zoomStep
+        if (this.options.initialPoint) {
+          this.map.setView(
+            [this.options.initialPoint.x, this.options.initialPoint.y],
+            mutation.payload.options.zoomLevel
+          );
+        } else {
+          this.map.setView(
+            this.startingPoint,
+            mutation.payload.options.zoomLevel
+          );
+        }
+        this.map.wheelPxPerZoomLevel = 200;
       }
     });
   },
   methods: {
-    initSettings() {
-      if (this.isSettings) return;
-      var ComponentClass = Vue.extend(dashMapUserSettings);
-      let test = new ComponentClass({
-        propsData: { idDashFrom: this.idDashFrom, idElement: this.idFrom },
-        vuetify,
-        store,
-      });
-      test.$mount();
-      let element = document.getElementsByClassName(
-        "leaflet-control-container"
-      );
-      let container = element[0];
-      container.appendChild(test.$el);
-      this.isSettings = true;
-    },
     reDrawMap(dataRest) {
       this.clearMap();
       this.error = null;
@@ -158,10 +116,37 @@ export default {
         this.createMap();
         //рисуем объекты на карте
         this.drawObjects(dataRest);
-        this.map.setView(this.startingPoint, this.options.zoomLevel);
+        if (this.options.initialPoint)
+          this.map.setView(
+            [this.options.initialPoint.x, this.options.initialPoint.y],
+            this.options.zoomLevel
+          );
+        else this.map.setView(this.startingPoint, this.options.zoomLevel);
         // this.clustering(dataRest);
       }
     },
+
+    initSettings() {
+      if (this.isSettings) return;
+      let ComponentClass = Vue.extend(dashMapUserSettings);
+      let test = new ComponentClass({
+        propsData: {
+          idDashFrom: this.idDashFrom,
+          idElement: this.idFrom,
+          map: this.map,
+        },
+        vuetify,
+        store,
+      });
+      test.$mount();
+      let element = document.getElementsByClassName(
+        "leaflet-control-container"
+      );
+      let container = element[0];
+      container.appendChild(test.$el);
+      this.isSettings = true;
+    },
+
     initTheme() {
       let options = this.$store.getters.getOptions({
         idDash: this.idDashFrom,
@@ -253,7 +238,9 @@ export default {
         idDash: this.idDashFrom,
         id: this.idFrom,
       });
-      if (options.osmserver) {
+      if (options.selectedLayer) {
+        this.osmserver = options.selectedLayer;
+      } else if (options.osmserver) {
         this.osmserver = options.osmserver;
       } else {
         this.error = "Введите osm server";
@@ -341,14 +328,19 @@ export default {
     },
 
     initMap() {
-
       this.map = L.map(this.$refs.map, {
-        wheelPxPerZoomLevel: this.options.zoomStep || 10,
+        wheelPxPerZoomLevel: 1 / this.options.zoomStep || 30,
         zoomSnap: 0,
         zoom: 10,
         maxZoom: 25,
       });
-
+      this.map.on("zoomend", () => {
+        let layers = document.getElementsByClassName("leaflet-marker-icon");
+        for (let x of layers) {
+          x.style.width = (x.naturalWidth / 10) * this.map.getZoom() + "px";
+          x.style.height = (x.naturalHeight / 10) * this.map.getZoom() + "px";
+        }
+      });
     },
 
     drawObjects(dataRest) {
@@ -359,7 +351,7 @@ export default {
           continue;
         }
         if (dataRest[i].ID === "1") {
-          const _point =  dataRest[i].coordinates.split(":");
+          const _point = dataRest[i].coordinates.split(":");
           const _coord = _point[1].split(",");
           this.startingPoint = [_coord[0], _coord[1]];
         }
@@ -397,7 +389,7 @@ export default {
       const _point = element.coordinates.split(":");
       const _coord = _point[1].split(",");
       this.startingPoint = [_coord[0], _coord[1]];
-      L.marker([_coord[0], _coord[1]], {
+      let marker = L.marker([_coord[0], _coord[1]], {
         icon: icon,
         zIndexOffset: -1000,
         riseOnHover: true,
@@ -408,6 +400,7 @@ export default {
           direction: "top",
           className: "leaftet-hover",
         });
+      L.DomUtil.addClass(marker._icon, "className");
     },
 
     drawMarkerHTML({ lib, element, isCenter }) {
@@ -582,14 +575,31 @@ export default {
       this.map.addLayer(cluster);
     },
     createMap() {
-      if (this.maptheme === "black") {
-        this.tileLayer = L.tileLayer.colorFilter(this.osmserver, {
-          filter: ["grayscale:100%", "invert:100%"],
-        });
+      let tileLayer;
+      if (!this.osmserver) return;
+      if (typeof this.osmserver === "string") {
+        if (this.maptheme === "black") {
+          tileLayer = L.tileLayer.colorFilter(this.osmserver, {
+            filter: ["grayscale:100%", "invert:100%"],
+          });
+        } else {
+          tileLayer = L.tileLayer.colorFilter(this.osmserver);
+        }
       } else {
-        this.tileLayer = L.tileLayer.colorFilter(this.osmserver);
+        if (!this.osmserver.tile) return;
+
+        let temp = this.osmserver.tile;
+        if (typeof this.osmserver.tile === "string") {
+          temp = [this.osmserver.tile];
+        }
+        if (this.maptheme === "black") {
+          temp[1].filter = ["grayscale:100%", "invert:100%"];
+          tileLayer = L.tileLayer.colorFilter(...temp);
+        } else {
+          tileLayer = L.tileLayer(...temp);
+        }
       }
-      this.tileLayer.addTo(this.map);
+      tileLayer.addTo(this.map);
     },
   },
 };
