@@ -7,16 +7,23 @@
         <v-dialog v-model="dialog" max-width="290">
           <template v-slot:activator="{ on, attrs }">
             <v-btn rounded color="#191919" v-bind="attrs" v-on="on">
-              <v-icon :style="{ color: theme.$title }">{{
-                mdiSettings
-              }}</v-icon>
+              <v-icon :style="{ color: theme.$title }">
+                {{ mdiSettings }}
+              </v-icon>
             </v-btn>
           </template>
-          <v-card >
+          <v-card>
             <v-card-title class="text-h5"> Настройки </v-card-title>
             <v-card-text>
               <p>Подложка</p>
-              <v-select :items="tileLayers" />
+              <v-select
+                v-model="options.selectedLayer"
+                return-object
+                :items="tileLayers"
+                item-text="name"
+                item-value="tile[0]"
+                @change="updateTileLayer($event)"
+              />
 
               <p>Начальный зум</p>
               <v-slider
@@ -41,8 +48,9 @@
               <v-slider
                 v-model="options.zoomStep"
                 class="align-center"
-                max="200"
-                min="1"
+                max="1"
+                min="0.01"
+                step="0.01"
               >
                 <template v-slot:label>
                   <v-text-field
@@ -55,6 +63,16 @@
                   />
                 </template>
               </v-slider>
+
+              <p>Начальная точка</p>
+              <v-row>
+                <v-col col="6">
+                  <v-text-field v-model="options.initialPoint.x" label="X" />
+                </v-col>
+                <v-col col="6">
+                  <v-text-field v-model="options.initialPoint.y" label="Y" />
+                </v-col>
+              </v-row>
 
               <p>Легенда карты</p>
               <v-checkbox
@@ -69,49 +87,128 @@
       </v-row>
 
       <v-row
+        v-show="options.showLegend"
         align="end"
         align-content="end"
-        v-show="options.showLegend"
         class="mb-5 mr-0"
       >
         <v-spacer></v-spacer>
         <v-card
           style="max-height: 466px"
           max-width="280"
-          class="pa-5"
+          class="px-5 pb-5"
           color="#191919"
         >
           <v-subheader style="color: white" class="px-0">
-            <v-icon :style="{ color: theme.$title }">{{ mdiList }}</v-icon>
-            <span
-              style="
-                font-family: Proxima Nova;
-                font-size: 18px;
-                font-style: normal;
-                font-weight: 600;
-                line-height: 22px;
-                letter-spacing: 0em;
-                text-align: left;
-              ">
-              Легенда
-            </span>
+            <v-row class="ma-0 fill-height"  >
+              <v-col class="ma-0 pa-0 fill-height">
+                <v-row class="mt-5 mx-0 pa-0" justify="center" align="center"> 
+                  <svg
+                    width="24"
+                    height="24"
+                    viewBox="0 0 24 24"
+                    fill="none"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <g clip-path="url(#clip0)">
+                      <path
+                        d="M19 5V19H5V5H19ZM20.1 3H3.9C3.4 3 3 3.4 3 3.9V20.1C3 20.5 3.4 21 3.9 21H20.1C20.5 21 21 20.5 21 20.1V3.9C21 3.4 20.5 3 20.1 3ZM11 7H17V9H11V7ZM11 11H17V13H11V11ZM11 15H17V17H11V15ZM7 7H9V9H7V7ZM7 11H9V13H7V11ZM7 15H9V17H7V15Z"
+                        fill="white"
+                      />
+                    </g>
+                    <defs>
+                      <clipPath id="clip0">
+                        <rect width="24" height="24" fill="white" />
+                      </clipPath>
+                    </defs>
+                  </svg>
+
+                  <span
+                    style="
+                  margin-left: 10px;
+                  font-family: Proxima Nova;
+                  font-size: 18px;
+                  font-style: normal;
+                  font-weight: 600;F
+                  line-height: 22px;
+                  letter-spacing: 0em;
+                  text-align: left;
+                "
+                  >
+                    Легенда
+                  </span>
+                  <v-spacer />
+                  <a style="align-self: center" @click="closeLegend">
+                    <svg
+                      width="20"
+                      height="20"
+                      viewBox="0 0 20 20"
+                      fill="none"
+                      xmlns="http://www.w3.org/2000/svg"
+                    >
+                      <path
+                        d="M15.8332 5.3415L14.6582 4.1665L9.99984 8.82484L5.3415 4.1665L4.1665 5.3415L8.82484 9.99984L4.1665 14.6582L5.3415 15.8332L9.99984 11.1748L14.6582 15.8332L15.8332 14.6582L11.1748 9.99984L15.8332 5.3415Z"
+                        fill="#DADADA"
+                      />
+                    </svg>
+                  </a>
+                </v-row>
+              </v-col>
+            </v-row>
           </v-subheader>
-          <v-divider></v-divider>
+          <v-divider
+            style="margin-bottom: 10px; border: 1px solid #555454"
+          ></v-divider>
+
           <v-card width="240" outlined color="white">
             <v-list style="max-height: 382px" class="overflow-y-auto">
-              <v-list-item
-                three-line
-                v-for="item in library.objects"
-                :key="item.name"
-                v-if="item.image"
-              >
-                <v-list-item-avatar>
-                  <v-img :src="base_svg_url + item.image"></v-img>
-                </v-list-item-avatar>
+              <v-list-item v-for="item in library.objects" :key="item.name">
+                <template v-if="item.image">
+                  <v-list-item-avatar size="20px" style="align-self: center">
+                    <v-img :src="base_svg_url + item.image"></v-img>
+                  </v-list-item-avatar>
 
-                <v-list-item-content>
-                  <v-list-item-title v-text="item.name"></v-list-item-title>
-                </v-list-item-content>
+                  <v-list-item-title
+                    style="text-align: left"
+                    v-text="item.name"
+                  ></v-list-item-title>
+                </template>
+
+                <template v-else-if="item.background_color">
+                  <v-list-item-avatar
+                    size="20px"
+                    v-html="createHtmlIcon(item)"
+                    style="align-self: center; border-radius: 0%"
+                  />
+
+                  <v-list-item-title
+                    style="text-align: left"
+                    v-text="item.name"
+                  />
+                </template>
+
+                <template v-else>
+                  <v-list-item-avatar size="20px" style="align-self: center">
+                    <div>
+                      <svg height="210" width="200">
+                        <line
+                          x1="0"
+                          y1="0"
+                          x2="200"
+                          y2="200"
+                          :stroke="item.color"
+                          :stroke-width="item.width"
+                        />
+                      </svg>
+                    </div>
+                  </v-list-item-avatar>
+                  <v-list-item-content>
+                    <v-list-item-title
+                      style="text-align: left"
+                      v-text="item.name"
+                    />
+                  </v-list-item-content>
+                </template>
               </v-list-item>
             </v-list>
           </v-card>
@@ -123,10 +220,16 @@
 
 <script>
 import { mdiFormatListBulletedSquare, mdiSettings } from "@mdi/js";
+import L from "leaflet";
+import "leaflet/dist/leaflet.css";
+import "leaflet.tilelayer.colorfilter";
+import "leaflet.markercluster";
 export default {
   props: {
     idElement: String,
     idDashFrom: String,
+    map: Object,
+    // library: Object
   },
   data() {
     return {
@@ -134,12 +237,54 @@ export default {
       mdiList: mdiFormatListBulletedSquare,
       dialog: false,
       base_svg_url: `${window.location.origin}/svg/`,
-      tileLayers: ["test", "test1"],
+      currentTile: {},
+      tileLayers: [
+        {
+          name: "Заданная в настройках",
+          tile: [],
+        },
+        // {
+        //   name: "Яндекс карта",
+        //   tile: [
+        //     "http://vec{s}.maps.yandex.net/tiles?l=map&v=4.55.2&z={z}&x={x}&y={y}&scale=2&lang=ru_RU",
+        //     {
+        //       subdomains: ["01", "02", "03", "04"],
+        //       attribution: '<a http="yandex.ru" target="_blank">Яндекс</a>',
+        //       reuseTiles: true,
+        //       updateWhenIdle: false,
+        //     },
+        //   ],
+        // },
+        {
+          name: "Google спутник",
+          tile: [
+            "http://{s}.google.com/vt/lyrs=s,h&x={x}&y={y}&z={z}",
+            {
+              subdomains: ["mt0", "mt1", "mt2", "mt3"],
+              attribution: '<a http="google.ru" target="_blank">Google</a>',
+            },
+          ],
+        },
+        {
+          name: "Google карты",
+          tile: [
+            "http://{s}.google.com/vt/lyrs=p&x={x}&y={y}&z={z}",
+            {
+              subdomains: ["mt0", "mt1", "mt2", "mt3"],
+              attribution: '<a http="google.ru" target="_blank">Google</a>',
+            },
+          ],
+        },
+      ],
       options: {
         selected: "яндекс",
-        selectedLayer: "test",
+        selectedLayer: "",
         zoomLevel: 10,
-        zoomStep: 25,
+        zoomStep: 0.01,
+        initialPoint: {
+          x: 59.242065955847735,
+          y: 74.35169122692963,
+        },
         showLegend: true,
       },
     };
@@ -171,26 +316,71 @@ export default {
       idDash: this.idDashFrom,
       id: this.idElement,
     });
-    
+    this.tileLayers[0].tile = options.osmserver;
     // init store for reactivity
-    if (!options.showLegend) {
+    if (!options.showLegend || !options.initialPoint) {
       let initOptions = {
         showLegend: true,
         zoomLevel: this.options.zoomLevel,
         zoomStep: this.options.zoomStep,
         selectedLayer: this.options.selectedLayer,
+        initialPoint: this.options.initialPoint,
       };
       this.$store.commit("setOptions", {
         idDash: this.idDashFrom,
         id: this.idElement,
         options: initOptions,
       });
-    } 
-    else {
-      this.options = options
+    } else {
+      this.options = options;
     }
+    this.setTileLayer();
   },
   methods: {
+    closeLegend() {
+      this.options.showLegend = false;
+    },
+    createHtmlIcon(lib) {
+      let {
+        text_color: textColor = "#FFFFFF",
+        background_color: color = "65, 62, 218",
+        opacity = 0.6,
+        label_field: text = "КП-240",
+        border_radius: borderRadius = "2px",
+        border = "none",
+        width = 20,
+        height = 20,
+      } = lib;
+      return `<div class="leaflet-div-icon" 
+          style="
+            background-color: ${color};
+            opacity: ${opacity};
+            mix-blend-mode: normal;
+            border: ${border};
+            border-radius: ${borderRadius}px;
+            padding: 2px 6px;
+            display: inline-block;
+            font-size: 14px;
+            font-weight: 600;
+        ">
+          <span style="color:${textColor}">кп<span>
+        </div>`;
+    },
+    updateTileLayer(e) {
+      this.map.removeLayer(this.currentTile);
+      if (typeof e.tile === "string") {
+        let temp = e.tile;
+        temp = [temp];
+        this.currentTile = L.tileLayer(...temp);
+        this.map.addLayer(this.currentTile);
+        this.updateOptions({ selectedLayer: temp[0] });
+        return;
+      }
+      this.currentTile = L.tileLayer(...e.tile);
+      this.map.addLayer(this.currentTile);
+      this.updateOptions({ selectedLayer: e.tile[0] });
+    },
+
     updateOptions(newOptions) {
       this.$store.commit("updateOptions", {
         idDash: this.idDashFrom,
@@ -198,6 +388,65 @@ export default {
         options: { ...this.dashSettings, ...newOptions },
       });
     },
+
+    setTileLayer() {
+      // this.map.removeLayer(grayscale);
+      // this.map.addLayer(streets);
+      // let test = L.tileLayer(
+      //   "http://vec{s}.maps.yandex.net/tiles?l=map&v=4.55.2&z={z}&x={x}&y={y}&scale=2&lang=ru_RU",
+      //   {
+      //     subdomains: ["01", "02", "03", "04"],
+      //     attribution: '<a http="yandex.ru" target="_blank">Яндекс</a>',
+      //     reuseTiles: true,
+      //     updateWhenIdle: false,
+      //   }
+      // ).addTo(this.map);
+      // this.map.addLayer(test)
+      // let baselayers = {
+      //   "Tile Layer 1": L.tileLayer(
+      //     "http://tile2.maps.2gis.com/tiles?x={x}&y={y}&z={z}"
+      //   ),
+      //   "Tile Layer 2": L.tileLayer(
+      //     "http://{s}.google.com/vt/lyrs=m&x={x}&y={y}&z={z}",
+      //     {
+      //       subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      //     }
+      //   ),
+      //   "Tile Layer 3": L.tileLayer(
+      //     "http://{s}.google.com/vt/lyrs=s,h&x={x}&y={y}&z={z}",
+      //     {
+      //       subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      //     }
+      //   ),
+      // };
+      // let overlays = {};
+      // baselayers["Tile Layer 1"].addTo(this.map);
+      // baselayers["Tile Layer 2"].addTo(this.map);
+      // L.control.layers(baselayers, overlays).addTo(this.map);
+      // L.tileLayer("http://{s}.google.com/vt/lyrs=s,h&x={x}&y={y}&z={z}", {
+      //   subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      // });
+      // L.tileLayer("http://{s}.google.com/vt/lyrs=m&x={x}&y={y}&z={z}", {
+      //   subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      // }).addTo(this.map);
+      // let terrain = L.tileLayer(
+      //   "http://{s}.google.com/vt/lyrs=p&x={x}&y={y}&z={z}",
+      //   {
+      //     subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      //   }
+      // ).addTo(this.map);
+      // let terrain1 = L.tileLayer(
+      //   "http://{s}.google.com/vt/lyrs=s,h&x={x}&y={y}&z={z}",
+      //   {
+      //     subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      //   }
+      // ).addTo(this.map);
+      // L.tileLayer("http://{s}.google.com/vt/lyrs=s&x={x}&y={y}&z={z}", {
+      //   subdomains: ["mt0", "mt1", "mt2", "mt3"],
+      // }).addTo(this.map);
+      // this.map.addLayer(terrain);
+    },
+
     setOptions: function () {
       // отправляем настройки в хранилище
       if (!this.options.level) {
